Add progress indicator for remote saving.
Quite an ordeal for such a simple thing. The problem is that I wanted
display a 'spinner', display different colors and texts for
success/failures and then I had to empty the notification area again.
Maybe a ui library could do it better, but I'd be more interested in the
abstraction it would use.

Probably it could be done better using rx, e.g. sequencing the three
behaviours, removing the previous when the next 'phase' occurred?

    interval progress (every 250ms) -> req success/failure -> cleaning (5s after)