Close that span element.
As it turns out, for void (formerly self-closing) elements `<el />`
means just `<el>` [1] and `span` isn't even a void element [2]. So for
my own sanity, close that span element and keep those `script` tags out
of it.  (Why did `.textContent` not "remove" those script elements? Or
did it? As it turns out, it did but everything in that file was
evaluated before it was removed and everyone was happy.)

[1]: http://dev.w3.org/html5/spec-author-view/syntax.html#syntax-start-tag
[2]: http://www.w3.org/html/wg/drafts/html/master/single-page.html#void-elements